+++ src/qibotn/eval_qu.py
@@ -18,11 +18,6 @@ def init_state_tn(nqubits, init_state_sv):
 
     return qtn.tensor_1d.MatrixProductState.from_dense(init_state_sv, dims)
 
-def init_state_tn_tebd(initial_state):
-
-    initial_state = qtn.MPS_computational_state(initial_state)
-    return initial_state
-
 def dense_vector_tn_qu(qasm: str, initial_state, mps_opts, backend="numpy"):
     """Evaluate circuit in QASM format with Quimb.
 
@@ -49,53 +44,3 @@ def dense_vector_tn_qu(qasm: str, initial_state, mps_opts, backend="numpy"):
     amplitudes = interim.to_dense(backend=backend)
 
     return amplitudes
-
-def tebd_tn_qu(circuit, tebd_opts):
-
-    from qibo.hamiltonians import SymbolicHamiltonian
-
-    hamiltonian = tebd_opts["hamiltonian"]
-    dt = tebd_opts["dt"]
-    initial_state = tebd_opts["initial_state"]
-    tot_time = tebd_opts["tot_time"]
-    nqubits = circuit.nqubits
-
-    init_state = init_state_tn_tebd(initial_state)
-    from qibo import hamiltonians
-
-    if hamiltonian == "TFIM":
-        ham = hamiltonians.TFIM(nqubits=nqubits, dense=False)
-    elif hamiltonian == "NIX":
-        ham = hamiltonians.X(nqubits=nqubits, dense=False)
-    elif hamiltonian == "NIY":
-        ham = hamiltonians.Y(nqubits=nqubits, dense=False)
-    elif hamiltonian == "NIZ":
-        ham = hamiltonians.Z(nqubits=nqubits, dense=False)
-    elif hamiltonian == "XXZ":
-        ham = hamiltonians.XXZ(nqubits=nqubits, dense=False)
-    elif hamiltonian == "MC":
-        ham = hamiltonians.MaxCut(nqubits=nqubits, dense=False)
-    else:
-        raise_error(
-            NotImplementedError, "QiboTN does not support custom hamiltonians"
-            )
-
-    terms_dict = {}
-    i=0
-    list_of_terms = ham.terms
-    for t in list_of_terms:
-        terms_dict.update({None: t.matrix})
-        i=i+1
-    
-    H = qtn.LocalHam1D(nqubits, H2=terms_dict)
-
-    tebd = qtn.TEBD(init_state, H)
-    ts = np.arange(0, tot_time, dt)
-
-    states = {}
-    for t in tebd.at_times(ts, tol=1e-3):
-        states.update({None: t.to_dense()}) 
-
-    state = np.array(list(states.values()))[-1]
-
-    return state
