+++ src/qibotn/working_as_independent_fn.py
@@ -1,5 +1,6 @@
 """ TEBD usage : An independent algorithm function"""
-import eval_qu as eval
+#import eval_qu as eval
+import tebd
 import qibo
 from qibo import hamiltonians
 
@@ -10,4 +11,4 @@ tebd_opts = {"dt":1e-4, "hamiltonian": "XXZ", "initial_state": "10101", "tot_tim
 ham = hamiltonians.XXZ(nqubits=5, dense=False)
 circuit= ham.circuit(dt=1e-4)
 
-print(eval.tebd_tn_qu(circuit, tebd_opts))
+print(tebd.tebd_quimb(circuit, tebd_opts))
